+++ transcribe.py
@@ -13,8 +13,10 @@ from resampy.core import resample
 from torchaudio.pipelines import MMS_FA as bundle
 from tqdm.auto import tqdm
 
-from corrector import BertModel, corrector
+from corrector import corrector
+from BertModel import BertModel
 from denoiser import denoiser
+from TranscribeResult import TranscribeResult
 from utils import load_dict
 
 logger = logging.getLogger(__name__)
@@ -40,23 +42,6 @@ aligner = bundle.get_aligner()
 SAMPLE_RATE = 16000
 
 t2s_char_dict, char_jyutping_dict, jyutping_char_dict, chars_freq = load_dict()
-
-
-class TranscribeResult:
-    """
-    Each TranscribeResult object represents one SRT line.
-    """
-
-    def __init__(self, text: str, start_time: float, end_time: float):
-        self.text = text
-        self.start_time = start_time
-        self.end_time = end_time
-
-    def __str__(self):
-        return f"TranscribeResult(text={self.text}, start_time={self.start_time}, end_time={self.end_time})"
-
-    def __repr__(self):
-        return str(self)
 
 
 def asr(
